Fixed many warnings.

+++ source/Core/Castor3D/Cache/BillboardCache.cpp
@@ -82,7 +82,7 @@ namespace castor3d
 					, entry.second.pass
 					, instanceMult );
 				auto it = m_entries.emplace( entry.second.hash, entry.second ).first;
-				it->second.id = m_entries.size();
+				it->second.id = int32_t( m_entries.size() );
 
 				if ( instanceMult )
 				{
@@ -258,7 +258,7 @@ namespace castor3d
 		{
 			auto & uboPools = *device.uboPools;
 			auto & baseEntry = iresult.first->second;
-			baseEntry.id = m_baseEntries.size();
+			baseEntry.id = int32_t( m_baseEntries.size() );
 			baseEntry.modelUbo = uboPools.getBuffer< ModelUboConfiguration >( VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT );
 			baseEntry.billboardUbo = uboPools.getBuffer< BillboardUboConfiguration >( VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT );
